requirements.txt
# requirements.txt for NLM-toolkit
# Standard library only for basic scripts

watchdog
pymupdf
nltk
spacy
tqdm
python-dotenv
markdown
weasyprint
chromadb
sentence-transformers
# The following are optional, only needed for advanced features:
faiss-cpu
# Ollama must be installed separately for local LLM summarization (https://ollama.com/)
gradio
llama-cloud-services